Moved numericUpDownHexFix to its own file
